Move anta.tools.misc to anta.logger
