+++ anta/logger.py
@@ -7,6 +7,7 @@ Configure logging for ANTA
 from __future__ import annotations
 
 import logging
+import traceback
 from enum import Enum
 from pathlib import Path
 from typing import Literal, Optional
@@ -14,7 +15,6 @@ from typing import Literal, Optional
 from rich.logging import RichHandler
 
 from anta import __DEBUG__
-from anta.tools.misc import exc_to_str
 
 logger = logging.getLogger(__name__)
 
@@ -87,6 +87,13 @@ def setup_logging(level: LogLevel = Log.INFO, file: Path | None = None) -> None:
         logger.debug("ANTA Debug Mode enabled")
 
 
+def exc_to_str(exception: BaseException) -> str:
+    """
+    Helper function that returns a human readable string from an BaseException object
+    """
+    return f"{type(exception).__name__}{f': {exception}' if str(exception) else ''}"
+
+
 def anta_log_exception(exception: BaseException, message: Optional[str] = None, calling_logger: Optional[logging.Logger] = None) -> None:
     """
     Helper function to help log exceptions:
@@ -94,7 +101,7 @@ def anta_log_exception(exception: BaseException, message: Optional[str] = None, 
     * otherwise logger.error is called
 
     Args:
-        exception (BAseException): The Exception being logged
+        exception (BaseException): The Exception being logged
         message (str): An optional message
         calling_logger (logging.Logger): A logger to which the exception should be logged
                                          if not present, the logger in this file is used.
@@ -105,3 +112,10 @@ def anta_log_exception(exception: BaseException, message: Optional[str] = None, 
     calling_logger.critical(f"{message}\n{exc_to_str(exception)}" if message else exc_to_str(exception))
     if __DEBUG__:
         calling_logger.exception(f"[ANTA Debug Mode]{f' {message}' if message else ''}", exc_info=exception)
+
+
+def tb_to_str(exception: BaseException) -> str:
+    """
+    Helper function that returns a traceback string from an BaseException object
+    """
+    return "Traceback (most recent call last):\n" + "".join(traceback.format_tb(exception.__traceback__))
